[0.5.0] glib based event dispatcher
feat: glib based event dispatcher
fix: improve thread safety in all available dispatchers
fix: assigning same Variant object to itself
test: move dispatchers creation to main thread in gtests
test: reduce output from gtests
build: improve gtest CMakeLists
build: new CMake build option: DISPATCHER_GLIBMM

+++ examples/01_trafficlight.cpp
@@ -50,7 +50,7 @@ public:
     {
         if (args.size() == 2)
         {
-            printf("----> onNextStateTransition (args=%ld, thread=%d, index=%d)\n", args.size(), (int)args[0].toInt64(), (int)args[1].toInt64());
+            printf("----> onNextStateTransition (args=%d, thread=%d, index=%d)\n", (int)args.size(), (int)args[0].toInt64(), (int)args[1].toInt64());
         }
         else
         {
